refactor: Car의 name 필드 wrapping
