+++ src/main/java/racingcar/domain/Car.java
@@ -1,19 +1,15 @@
 package racingcar.domain;
 
 public class Car {
-    private static final String LENGTH_ERROR_MESSAGE = "자동차 이름은 5글자 이하만 가능합니다.";
-    private static final int MAX_NAME_LENGTH = 5;
-    private final String name;
+    private final Name name;
     private int position = 0;
 
-
     private Car(String name) {
-        validateLength(name);
-        this.name = name;
+        this.name = new Name(name);
     }
 
     public String getName() {
-        return name;
+        return name.getName();
     }
 
     public int getPosition() {
@@ -27,12 +23,6 @@ public class Car {
     public void move(boolean isCanMove) {
         if (isCanMove) {
             position += 1;
-        }
-    }
-
-    private void validateLength(String name) {
-        if (name.length() > MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException(LENGTH_ERROR_MESSAGE);
         }
     }
 }
